chore(app): add better-auth handlers
