+++ apps/app/src/lib/auth.functions.ts
@@ -1,0 +1,25 @@
+import { createServerFn } from "@tanstack/react-start";
+import { getRequestHeaders } from "@tanstack/react-start/server";
+
+import { auth } from "@repo/auth/server/tanstack";
+
+export const getSession = createServerFn({ method: "GET" }).handler(
+  async () => {
+    const headers = getRequestHeaders();
+
+    return auth.api.getSession({ headers });
+  },
+);
+
+export const ensureSession = createServerFn({ method: "GET" }).handler(
+  async () => {
+    const headers = getRequestHeaders();
+    const session = await auth.api.getSession({ headers });
+
+    if (!session) {
+      throw new Error("Unauthorized");
+    }
+
+    return session;
+  },
+);
