+++ apps/app/src/routes/api/auth/$.ts
@@ -1,0 +1,12 @@
+import { createFileRoute } from "@tanstack/react-router";
+
+import { auth } from "@repo/auth/server/tanstack";
+
+export const Route = createFileRoute("/api/auth/$")({
+  server: {
+    handlers: {
+      GET: ({ request }) => auth.handler(request),
+      POST: ({ request }) => auth.handler(request),
+    },
+  },
+});
