+++ packages/auth/package.json
@@ -5,6 +5,7 @@
   "exports": {
     "./client": "./src/client.ts",
     "./server": "./src/server.ts",
+    "./server/tanstack": "./src/server/tanstack.ts",
     "./server/mcp": "./src/server/mcp.ts"
   },
   "scripts": {
